Remove unnecessary 2GB heap and add memory profiling
- Remove 2GB heap requirement for integration tests (512MB is sufficient)
- Add memory profiling test to measure buffer usage
- Add platform-specific memory helpers for GC and heap measurement
- Split monolithic AutobahnTests into focused test classes
- Profiling shows ~2MB delta for 10x 1MB payloads, confirming
  memory usage is reasonable without extra heap

Test failures in CI are race conditions (ClosedChannelException),
not memory issues.
