+++ .docker/config/fuzzingserver.json
@@ -2,9 +2,6 @@
   "url": "ws://127.0.0.1:9001",
   "outdir": "./reports/clients",
   "cases": ["*"],
-  "exclude-cases": [
-    "12.*",
-    "13.*"
-  ],
+  "exclude-cases": [],
   "exclude-agent-cases": {}
 }
