+++ src/commonTest/kotlin/com/ditchoom/websocket/AutobahnCase10MiscTests.kt
@@ -1,0 +1,11 @@
+package com.ditchoom.websocket
+
+import kotlin.test.Test
+
+class AutobahnCase10MiscTests {
+    @Test
+    fun case10_1_1() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(301)
+        }
+}
