+++ src/commonTest/kotlin/com/ditchoom/websocket/AutobahnCase12CompressionTests.kt
@@ -1,0 +1,275 @@
+package com.ditchoom.websocket
+
+import kotlin.test.Test
+
+class AutobahnCase12CompressionTests {
+    @Test
+    fun case12_1_1() = runTestNoTimeSkipping { echoMessageAndClose(302, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_1_2() = runTestNoTimeSkipping { echoMessageAndClose(303, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_1_3() = runTestNoTimeSkipping { echoMessageAndClose(304, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_1_4() = runTestNoTimeSkipping { echoMessageAndClose(305, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_1_5() = runTestNoTimeSkipping { echoMessageAndClose(306, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_1_6() = runTestNoTimeSkipping { echoMessageAndClose(307, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_1_7() = runTestNoTimeSkipping { echoMessageAndClose(308, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_1_8() = runTestNoTimeSkipping { echoMessageAndClose(309, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_1_9() = runTestNoTimeSkipping { echoMessageAndClose(310, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_1_10() = runTestNoTimeSkipping { echoMessageAndClose(311, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_1_11() = runTestNoTimeSkipping { echoMessageAndClose(312, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_1_12() = runTestNoTimeSkipping { echoMessageAndClose(313, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_1_13() = runTestNoTimeSkipping { echoMessageAndClose(314, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_1_14() = runTestNoTimeSkipping { echoMessageAndClose(315, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_1_15() = runTestNoTimeSkipping { echoMessageAndClose(316, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_1_16() = runTestNoTimeSkipping { echoMessageAndClose(317, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_1_17() = runTestNoTimeSkipping { echoMessageAndClose(318, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_1_18() = runTestNoTimeSkipping { echoMessageAndClose(319, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_2_1() = runTestNoTimeSkipping { echoBinaryMessageAndClose(320, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_2_2() = runTestNoTimeSkipping { echoBinaryMessageAndClose(321, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_2_3() = runTestNoTimeSkipping { echoBinaryMessageAndClose(322, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_2_4() = runTestNoTimeSkipping { echoBinaryMessageAndClose(323, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_2_5() = runTestNoTimeSkipping { echoBinaryMessageAndClose(324, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_2_6() = runTestNoTimeSkipping { echoBinaryMessageAndClose(325, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_2_7() = runTestNoTimeSkipping { echoBinaryMessageAndClose(326, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_2_8() = runTestNoTimeSkipping { echoBinaryMessageAndClose(327, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_2_9() = runTestNoTimeSkipping { echoBinaryMessageAndClose(328, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_2_10() = runTestNoTimeSkipping { echoBinaryMessageAndClose(329, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_2_11() = runTestNoTimeSkipping { echoBinaryMessageAndClose(330, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_2_12() = runTestNoTimeSkipping { echoBinaryMessageAndClose(331, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_2_13() = runTestNoTimeSkipping { echoBinaryMessageAndClose(332, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_2_14() = runTestNoTimeSkipping { echoBinaryMessageAndClose(333, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_2_15() = runTestNoTimeSkipping { echoBinaryMessageAndClose(334, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_2_16() = runTestNoTimeSkipping { echoBinaryMessageAndClose(335, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_2_17() = runTestNoTimeSkipping { echoBinaryMessageAndClose(336, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_2_18() = runTestNoTimeSkipping { echoBinaryMessageAndClose(337, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_3_1() = runTestNoTimeSkipping { echoBinaryMessageAndClose(338, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_3_2() = runTestNoTimeSkipping { echoBinaryMessageAndClose(339, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_3_3() = runTestNoTimeSkipping { echoBinaryMessageAndClose(340, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_3_4() = runTestNoTimeSkipping { echoBinaryMessageAndClose(341, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_3_5() = runTestNoTimeSkipping { echoBinaryMessageAndClose(342, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_3_6() = runTestNoTimeSkipping { echoBinaryMessageAndClose(343, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_3_7() = runTestNoTimeSkipping { echoBinaryMessageAndClose(344, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_3_8() = runTestNoTimeSkipping { echoBinaryMessageAndClose(345, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_3_9() = runTestNoTimeSkipping { echoBinaryMessageAndClose(346, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_3_10() = runTestNoTimeSkipping { echoBinaryMessageAndClose(347, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_3_11() = runTestNoTimeSkipping { echoBinaryMessageAndClose(348, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_3_12() = runTestNoTimeSkipping { echoBinaryMessageAndClose(349, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_3_13() = runTestNoTimeSkipping { echoBinaryMessageAndClose(350, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_3_14() = runTestNoTimeSkipping { echoBinaryMessageAndClose(351, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_3_15() = runTestNoTimeSkipping { echoBinaryMessageAndClose(352, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_3_16() = runTestNoTimeSkipping { echoBinaryMessageAndClose(353, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_3_17() = runTestNoTimeSkipping { echoBinaryMessageAndClose(354, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_3_18() = runTestNoTimeSkipping { echoBinaryMessageAndClose(355, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_4_1() = runTestNoTimeSkipping { echoMessageAndClose(356, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_4_2() = runTestNoTimeSkipping { echoMessageAndClose(357, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_4_3() = runTestNoTimeSkipping { echoMessageAndClose(358, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_4_4() = runTestNoTimeSkipping { echoMessageAndClose(359, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_4_5() = runTestNoTimeSkipping { echoMessageAndClose(360, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_4_6() = runTestNoTimeSkipping { echoMessageAndClose(361, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_4_7() = runTestNoTimeSkipping { echoMessageAndClose(362, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_4_8() = runTestNoTimeSkipping { echoMessageAndClose(363, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_4_9() = runTestNoTimeSkipping { echoMessageAndClose(364, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_4_10() = runTestNoTimeSkipping { echoMessageAndClose(365, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_4_11() = runTestNoTimeSkipping { echoMessageAndClose(366, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_4_12() = runTestNoTimeSkipping { echoMessageAndClose(367, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_4_13() = runTestNoTimeSkipping { echoMessageAndClose(368, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_4_14() = runTestNoTimeSkipping { echoMessageAndClose(369, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_4_15() = runTestNoTimeSkipping { echoMessageAndClose(370, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_4_16() = runTestNoTimeSkipping { echoMessageAndClose(371, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_4_17() = runTestNoTimeSkipping { echoMessageAndClose(372, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_4_18() = runTestNoTimeSkipping { echoMessageAndClose(373, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_5_1() = runTestNoTimeSkipping { echoBinaryMessageAndClose(374, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_5_2() = runTestNoTimeSkipping { echoBinaryMessageAndClose(375, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_5_3() = runTestNoTimeSkipping { echoBinaryMessageAndClose(376, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_5_4() = runTestNoTimeSkipping { echoBinaryMessageAndClose(377, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_5_5() = runTestNoTimeSkipping { echoBinaryMessageAndClose(378, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_5_6() = runTestNoTimeSkipping { echoBinaryMessageAndClose(379, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_5_7() = runTestNoTimeSkipping { echoBinaryMessageAndClose(380, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_5_8() = runTestNoTimeSkipping { echoBinaryMessageAndClose(381, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_5_9() = runTestNoTimeSkipping { echoBinaryMessageAndClose(382, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_5_10() = runTestNoTimeSkipping { echoBinaryMessageAndClose(383, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_5_11() = runTestNoTimeSkipping { echoBinaryMessageAndClose(384, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_5_12() = runTestNoTimeSkipping { echoBinaryMessageAndClose(385, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_5_13() = runTestNoTimeSkipping { echoBinaryMessageAndClose(386, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_5_14() = runTestNoTimeSkipping { echoBinaryMessageAndClose(387, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_5_15() = runTestNoTimeSkipping { echoBinaryMessageAndClose(388, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_5_16() = runTestNoTimeSkipping { echoBinaryMessageAndClose(389, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_5_17() = runTestNoTimeSkipping { echoBinaryMessageAndClose(390, 1000, requestCompression = true) }
+
+    @Test
+    fun case12_5_18() = runTestNoTimeSkipping { echoBinaryMessageAndClose(391, 1000, requestCompression = true) }
+}
