+++ src/commonTest/kotlin/com/ditchoom/websocket/AutobahnCase2FragmentationTests.kt
@@ -1,0 +1,71 @@
+package com.ditchoom.websocket
+
+import kotlin.test.Test
+
+class AutobahnCase2FragmentationTests {
+    @Test
+    fun case2_1() =
+        runTestNoTimeSkipping {
+            prepareConnection(17)
+        }
+
+    @Test
+    fun case2_2() =
+        runTestNoTimeSkipping {
+            prepareConnection(18)
+        }
+
+    @Test
+    fun case2_3() =
+        runTestNoTimeSkipping {
+            prepareConnection(19)
+        }
+
+    @Test
+    fun case2_4() =
+        runTestNoTimeSkipping {
+            prepareConnection(20)
+        }
+
+    @Test
+    fun case2_5() =
+        runTestNoTimeSkipping {
+            prepareConnection(21)
+        }
+
+    @Test
+    fun case2_6() =
+        runTestNoTimeSkipping {
+            prepareConnection(22)
+        }
+
+    @Test
+    fun case2_7() =
+        runTestNoTimeSkipping {
+            prepareConnection(23)
+        }
+
+    @Test
+    fun case2_8() =
+        runTestNoTimeSkipping {
+            prepareConnection(24)
+        }
+
+    @Test
+    fun case2_9() =
+        runTestNoTimeSkipping {
+            prepareConnection(25)
+        }
+
+    @Test
+    fun case2_10() =
+        runTestNoTimeSkipping {
+            prepareConnection(26)
+        }
+
+    @Test
+    fun case2_11() =
+        runTestNoTimeSkipping {
+            prepareConnection(27)
+        }
+}
