+++ src/commonTest/kotlin/com/ditchoom/websocket/AutobahnCase6PingPongTests.kt
@@ -1,0 +1,875 @@
+package com.ditchoom.websocket
+
+import kotlin.test.Test
+
+class AutobahnCase6PingPongTests {
+    @Test
+    fun case6_1_1() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(65)
+        }
+
+    @Test
+    fun case6_1_2() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(66)
+        }
+
+    @Test
+    fun case6_1_3() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(67)
+        }
+
+    @Test
+    fun case6_2_1() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(68)
+        }
+
+    @Test
+    fun case6_2_2() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(69)
+        }
+
+    @Test
+    fun case6_2_3() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(70)
+        }
+
+    @Test
+    fun case6_2_4() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(71)
+        }
+
+    @Test
+    fun case6_3_1() =
+        runTestNoTimeSkipping {
+            prepareConnection(72)
+        }
+
+    @Test
+    fun case6_3_2() =
+        runTestNoTimeSkipping {
+            prepareConnection(73)
+        }
+
+    @Test
+    fun case6_4_1() =
+        runTestNoTimeSkipping {
+            prepareConnection(74)
+        }
+
+    @Test
+    fun case6_4_2() =
+        runTestNoTimeSkipping {
+            prepareConnection(75)
+        }
+
+    @Test
+    fun case6_4_3() =
+        runTestNoTimeSkipping {
+            prepareConnection(76)
+        }
+
+    @Test
+    fun case6_4_4() =
+        runTestNoTimeSkipping {
+            prepareConnection(77)
+        }
+
+    @Test
+    fun case6_5_1() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(78)
+        }
+
+    @Test
+    fun case6_5_2() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(79)
+        }
+
+    @Test
+    fun case6_5_3() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(80)
+        }
+
+    @Test
+    fun case6_5_4() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(81)
+        }
+
+    @Test
+    fun case6_5_5() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(82)
+        }
+
+    @Test
+    fun case6_6_1() =
+        runTestNoTimeSkipping {
+            prepareConnection(83)
+        }
+
+    @Test
+    fun case6_6_2() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(84)
+        }
+
+    @Test
+    fun case6_6_3() =
+        runTestNoTimeSkipping {
+            prepareConnection(85)
+        }
+
+    @Test
+    fun case6_6_4() =
+        runTestNoTimeSkipping {
+            prepareConnection(86)
+        }
+
+    @Test
+    fun case6_6_5() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(87)
+        }
+
+    @Test
+    fun case6_6_6() =
+        runTestNoTimeSkipping {
+            prepareConnection(88)
+        }
+
+    @Test
+    fun case6_6_7() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(89)
+        }
+
+    @Test
+    fun case6_6_8() =
+        runTestNoTimeSkipping {
+            prepareConnection(90)
+        }
+
+    @Test
+    fun case6_6_9() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(91)
+        }
+
+    @Test
+    fun case6_6_10() =
+        runTestNoTimeSkipping {
+            prepareConnection(92)
+        }
+
+    @Test
+    fun case6_6_11() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(93)
+        }
+
+    @Test
+    fun case6_7_1() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(94)
+        }
+
+    @Test
+    fun case6_7_2() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(95)
+        }
+
+    @Test
+    fun case6_7_3() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(96)
+        }
+
+    @Test
+    fun case6_7_4() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(97)
+        }
+
+    @Test
+    fun case6_8_1() =
+        runTestNoTimeSkipping {
+            prepareConnection(98)
+        }
+
+    @Test
+    fun case6_8_2() =
+        runTestNoTimeSkipping {
+            prepareConnection(99)
+        }
+
+    @Test
+    fun case6_9_1() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(100)
+        }
+
+    @Test
+    fun case6_9_2() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(101)
+        }
+
+    @Test
+    fun case6_9_3() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(102)
+        }
+
+    @Test
+    fun case6_9_4() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(103)
+        }
+
+    @Test
+    fun case6_10_1() =
+        runTestNoTimeSkipping {
+            prepareConnection(104)
+        }
+
+    @Test
+    fun case6_10_2() =
+        runTestNoTimeSkipping {
+            prepareConnection(105)
+        }
+
+    @Test
+    fun case6_10_3() =
+        runTestNoTimeSkipping {
+            prepareConnection(106)
+        }
+
+    @Test
+    fun case6_11_1() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(107)
+        }
+
+    @Test
+    fun case6_11_2() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(108)
+        }
+
+    @Test
+    fun case6_11_3() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(109)
+        }
+
+    @Test
+    fun case6_11_4() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(110)
+        }
+
+    @Test
+    fun case6_11_5() =
+        runTestNoTimeSkipping {
+            prepareConnection(111)
+        }
+
+    @Test
+    fun case6_12_1() =
+        runTestNoTimeSkipping {
+            prepareConnection(112)
+        }
+
+    @Test
+    fun case6_12_2() =
+        runTestNoTimeSkipping {
+            prepareConnection(113)
+        }
+
+    @Test
+    fun case6_12_3() =
+        runTestNoTimeSkipping {
+            prepareConnection(114)
+        }
+
+    @Test
+    fun case6_12_4() =
+        runTestNoTimeSkipping {
+            prepareConnection(115)
+        }
+
+    @Test
+    fun case6_12_5() =
+        runTestNoTimeSkipping {
+            prepareConnection(116)
+        }
+
+    @Test
+    fun case6_12_6() =
+        runTestNoTimeSkipping {
+            prepareConnection(117)
+        }
+
+    @Test
+    fun case6_12_7() =
+        runTestNoTimeSkipping {
+            prepareConnection(118)
+        }
+
+    @Test
+    fun case6_12_8() =
+        runTestNoTimeSkipping {
+            prepareConnection(119)
+        }
+
+    @Test
+    fun case6_13_1() =
+        runTestNoTimeSkipping {
+            prepareConnection(120)
+        }
+
+    @Test
+    fun case6_13_2() =
+        runTestNoTimeSkipping {
+            prepareConnection(121)
+        }
+
+    @Test
+    fun case6_13_3() =
+        runTestNoTimeSkipping {
+            prepareConnection(122)
+        }
+
+    @Test
+    fun case6_13_4() =
+        runTestNoTimeSkipping {
+            prepareConnection(123)
+        }
+
+    @Test
+    fun case6_13_5() =
+        runTestNoTimeSkipping {
+            prepareConnection(124)
+        }
+
+    @Test
+    fun case6_14_1() =
+        runTestNoTimeSkipping {
+            prepareConnection(125)
+        }
+
+    @Test
+    fun case6_14_2() =
+        runTestNoTimeSkipping {
+            prepareConnection(126)
+        }
+
+    @Test
+    fun case6_14_3() =
+        runTestNoTimeSkipping {
+            prepareConnection(127)
+        }
+
+    @Test
+    fun case6_14_4() =
+        runTestNoTimeSkipping {
+            prepareConnection(128)
+        }
+
+    @Test
+    fun case6_14_5() =
+        runTestNoTimeSkipping {
+            prepareConnection(129)
+        }
+
+    @Test
+    fun case6_14_6() =
+        runTestNoTimeSkipping {
+            prepareConnection(130)
+        }
+
+    @Test
+    fun case6_14_7() =
+        runTestNoTimeSkipping {
+            prepareConnection(131)
+        }
+
+    @Test
+    fun case6_14_8() =
+        runTestNoTimeSkipping {
+            prepareConnection(132)
+        }
+
+    @Test
+    fun case6_14_9() =
+        runTestNoTimeSkipping {
+            prepareConnection(133)
+        }
+
+    @Test
+    fun case6_14_10() =
+        runTestNoTimeSkipping {
+            prepareConnection(134)
+        }
+
+    @Test
+    fun case6_15_1() =
+        runTestNoTimeSkipping {
+            prepareConnection(135)
+        }
+
+    @Test
+    fun case6_16_1() =
+        runTestNoTimeSkipping {
+            prepareConnection(136)
+        }
+
+    @Test
+    fun case6_16_2() =
+        runTestNoTimeSkipping {
+            prepareConnection(137)
+        }
+
+    @Test
+    fun case6_16_3() =
+        runTestNoTimeSkipping {
+            prepareConnection(138)
+        }
+
+    @Test
+    fun case6_17_1() =
+        runTestNoTimeSkipping {
+            prepareConnection(139)
+        }
+
+    @Test
+    fun case6_17_2() =
+        runTestNoTimeSkipping {
+            prepareConnection(140)
+        }
+
+    @Test
+    fun case6_17_3() =
+        runTestNoTimeSkipping {
+            prepareConnection(141)
+        }
+
+    @Test
+    fun case6_17_4() =
+        runTestNoTimeSkipping {
+            prepareConnection(142)
+        }
+
+    @Test
+    fun case6_17_5() =
+        runTestNoTimeSkipping {
+            prepareConnection(143)
+        }
+
+    @Test
+    fun case6_18_1() =
+        runTestNoTimeSkipping {
+            prepareConnection(144)
+        }
+
+    @Test
+    fun case6_18_2() =
+        runTestNoTimeSkipping {
+            prepareConnection(145)
+        }
+
+    @Test
+    fun case6_18_3() =
+        runTestNoTimeSkipping {
+            prepareConnection(146)
+        }
+
+    @Test
+    fun case6_18_4() =
+        runTestNoTimeSkipping {
+            prepareConnection(147)
+        }
+
+    @Test
+    fun case6_18_5() =
+        runTestNoTimeSkipping {
+            prepareConnection(148)
+        }
+
+    @Test
+    fun case6_19_1() =
+        runTestNoTimeSkipping {
+            prepareConnection(149)
+        }
+
+    @Test
+    fun case6_19_2() =
+        runTestNoTimeSkipping {
+            prepareConnection(150)
+        }
+
+    @Test
+    fun case6_19_3() =
+        runTestNoTimeSkipping {
+            prepareConnection(151)
+        }
+
+    @Test
+    fun case6_19_4() =
+        runTestNoTimeSkipping {
+            prepareConnection(152)
+        }
+
+    @Test
+    fun case6_19_5() =
+        runTestNoTimeSkipping {
+            prepareConnection(153)
+        }
+
+    @Test
+    fun case6_20_1() =
+        runTestNoTimeSkipping {
+            prepareConnection(154)
+        }
+
+    @Test
+    fun case6_20_2() =
+        runTestNoTimeSkipping {
+            prepareConnection(155)
+        }
+
+    @Test
+    fun case6_20_3() =
+        runTestNoTimeSkipping {
+            prepareConnection(156)
+        }
+
+    @Test
+    fun case6_20_4() =
+        runTestNoTimeSkipping {
+            prepareConnection(157)
+        }
+
+    @Test
+    fun case6_20_5() =
+        runTestNoTimeSkipping {
+            prepareConnection(158)
+        }
+
+    @Test
+    fun case6_20_6() =
+        runTestNoTimeSkipping {
+            prepareConnection(159)
+        }
+
+    @Test
+    fun case6_20_7() =
+        runTestNoTimeSkipping {
+            prepareConnection(160)
+        }
+
+    @Test
+    fun case6_21_1() =
+        runTestNoTimeSkipping {
+            prepareConnection(161)
+        }
+
+    @Test
+    fun case6_21_2() =
+        runTestNoTimeSkipping {
+            prepareConnection(162)
+        }
+
+    @Test
+    fun case6_21_3() =
+        runTestNoTimeSkipping {
+            prepareConnection(163)
+        }
+
+    @Test
+    fun case6_21_4() =
+        runTestNoTimeSkipping {
+            prepareConnection(164)
+        }
+
+    @Test
+    fun case6_21_5() =
+        runTestNoTimeSkipping {
+            prepareConnection(165)
+        }
+
+    @Test
+    fun case6_21_6() =
+        runTestNoTimeSkipping {
+            prepareConnection(166)
+        }
+
+    @Test
+    fun case6_21_7() =
+        runTestNoTimeSkipping {
+            prepareConnection(167)
+        }
+
+    @Test
+    fun case6_21_8() =
+        runTestNoTimeSkipping {
+            prepareConnection(168)
+        }
+
+    @Test
+    fun case6_22_1() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(169)
+        }
+
+    @Test
+    fun case6_22_2() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(170)
+        }
+
+    @Test
+    fun case6_22_3() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(171)
+        }
+
+    @Test
+    fun case6_22_4() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(172)
+        }
+
+    @Test
+    fun case6_22_5() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(173)
+        }
+
+    @Test
+    fun case6_22_6() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(174)
+        }
+
+    @Test
+    fun case6_22_7() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(175)
+        }
+
+    @Test
+    fun case6_22_8() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(176)
+        }
+
+    @Test
+    fun case6_22_9() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(177)
+        }
+
+    @Test
+    fun case6_22_10() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(178)
+        }
+
+    @Test
+    fun case6_22_11() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(179)
+        }
+
+    @Test
+    fun case6_22_12() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(180)
+        }
+
+    @Test
+    fun case6_22_13() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(181)
+        }
+
+    @Test
+    fun case6_22_14() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(182)
+        }
+
+    @Test
+    fun case6_22_15() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(183)
+        }
+
+    @Test
+    fun case6_22_16() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(184)
+        }
+
+    @Test
+    fun case6_22_17() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(185)
+        }
+
+    @Test
+    fun case6_22_18() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(186)
+        }
+
+    @Test
+    fun case6_22_19() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(187)
+        }
+
+    @Test
+    fun case6_22_20() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(188)
+        }
+
+    @Test
+    fun case6_22_21() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(189)
+        }
+
+    @Test
+    fun case6_22_22() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(190)
+        }
+
+    @Test
+    fun case6_22_23() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(191)
+        }
+
+    @Test
+    fun case6_22_24() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(192)
+        }
+
+    @Test
+    fun case6_22_25() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(193)
+        }
+
+    @Test
+    fun case6_22_26() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(194)
+        }
+
+    @Test
+    fun case6_22_27() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(195)
+        }
+
+    @Test
+    fun case6_22_28() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(196)
+        }
+
+    @Test
+    fun case6_22_29() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(197)
+        }
+
+    @Test
+    fun case6_22_30() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(198)
+        }
+
+    @Test
+    fun case6_22_31() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(199)
+        }
+
+    @Test
+    fun case6_22_32() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(200)
+        }
+
+    @Test
+    fun case6_22_33() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(201)
+        }
+
+    @Test
+    fun case6_22_34() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(202)
+        }
+
+    @Test
+    fun case6_23_1() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(203)
+        }
+
+    @Test
+    fun case6_23_2() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(204)
+        }
+
+    @Test
+    fun case6_23_3() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(205)
+        }
+
+    @Test
+    fun case6_23_4() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(206)
+        }
+
+    @Test
+    fun case6_23_5() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(207)
+        }
+
+    @Test
+    fun case6_23_6() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(208)
+        }
+
+    @Test
+    fun case6_23_7() =
+        runTestNoTimeSkipping {
+            echoMessageAndClose(209)
+        }
+}
